WIP API AR implementation

+++ app/src/main/java/ml/moneo/app/model/Manual.kt
@@ -28,10 +28,10 @@ data class Overlay(
 
 data class Interaction(
     val id: String,
-    val x: Float,
-    val y: Float,
-    val width: Float,
-    val height: Float,
+    val x: Double,
+    val y: Double,
+    val width: Double,
+    val height: Double,
     val text: String,
     val overlay: Overlay,
 //    val overlayId: String,
